requirements.txt
blinker
certifi
charset-normalizer
click
cryptography
exceptiongroup
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Login
Flask-SQLAlchemy
gunicorn
idna
iniconfig
install
itsdangerous
Jinja2
MarkupSafe
packaging
pluggy
PyJWT
PyMySQL
python-dotenv
requests
six
SQLAlchemy
tomli
typing_extensions
urllib3
Werkzeug